🚚 remove getLatestChangelogDate function and related logic from InfoMenuList component
